PR Review fixes & more
