+++ src/config/config.js
@@ -1,6 +1,6 @@
 export const version ={release: 1, name:'1.0.0beta'};
 export const chainList=[
-    { short: 'BTS', name: 'BitShares'}
+    { short: 'BTS', name: 'BitShares', chainId: '4018d7844c78f6a6c41c6a552b898022310fc5dec06da467ee7905a8dad512c8'}
 ];
 export const nodeList = [
