+++ src/index.html
@@ -16,7 +16,7 @@
     <div class="main" id="app">
         <router-view name="header"></router-view>
         <router-view></router-view>
-        <Popups></Popups>
+        <Popups ref="popupComp"></Popups>
     </div>
     <script>
         import Vue from 'vue/dist/vue.js';
